requirements.txt
paho-mqtt
influxdb 
datetime
configparser
